+++ opt_purchase/data/record_rules.xml
@@ -11,6 +11,7 @@
             <field name="domain_force">[('type', '!=', 'in_invoice')]</field>
             <field eval="[(4,ref('purchase.group_purchase_user'))]" name="groups"/>
         </record>
+
         <record model="ir.rule" id="purchase_manager_vendor_bills">
             <field name="name">Allow Vendor Bills access to Purchase Managers</field>
             <field name="model_id" ref="model_account_invoice"/>
