+++ opt_purchase/models/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from . import purchase_order
-from . import ap_gl_account
 from . import account_move
+from . import ap_gl_account
 from . import expense_class
+from . import purchase_approval
+from . import purchase_order
