+++ opt_purchase/views/account_move_views.xml
@@ -14,7 +14,7 @@
                     <button name="action_invoice_to_approve" type="object" states="draft" string="To Approve" class="oe_highlight" groups="account.group_account_invoice"/>
                 </xpath>
                 <xpath expr="//field[@name='state']" position="attributes">
-                    <attribute name="statusbar_visible">draft,to_approve,open,paid</attribute>
+                    <attribute name="statusbar_visible">draft,to_approve,posted,paid</attribute>
                 </xpath>
                 <xpath expr="//field[@name='partner_bank_id']" position="after">
                     <field name="charge_code_id"/>
