+++ opt_purchase/views/purchase_order_views.xml
@@ -12,21 +12,8 @@
       	  <field name="can_edit_approval" invisible="1"/>
           <field name="ready_approval"/>
       	  <field name="approved" attrs="{'readonly': ['|', ('can_edit_approval', '=', False), ('ready_approval', '=', False)]}"/>
-
-<!--          <field name="date_approved"/>-->
-
       	</tree>
       </field>
-    </record>
-
-    <record id="action_purchase_approval_tree" model="ir.actions.act_window">
-      <field name="name">Purchase Approval</field>
-      <field name="res_model">purchase.approval</field>
-<!--      <field name="view_type">form</field>-->
-      <field name="view_mode">tree</field>
-      <field name="target">fullscreen</field>
-<!--      <field name="nodestroy" eval="False"/>-->
-      <field name="context">{}</field>
     </record>
 
     <record id="view_purchase_order_form_view_inherit_opt" model="ir.ui.view">
@@ -48,9 +35,6 @@
       	  <field name="ap_gl_account"/>
       	  <field name="expense_class"/>
       	</xpath>
-<!--      	<xpath expr="//button[@name='action_rfq_send']" position="attributes">-->
-<!--      	  <attribute name="attrs">{'invisible': ['|', ('approved', '=', False)]}</attribute>-->
-<!--      	</xpath>-->
       	<xpath expr="//button[@name='button_approve']" position="replace">
       	  <field name="show_action_approve" invisible="1"/>
       	  <button name="action_approve" string="Approve" type="object" class="oe_highlight" attrs="{'invisible': [('show_action_approve', '=', False)]}"/>
