+++ opt_purchase/wizard/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 from . import wizard_import_helper
-from . import account_invoice_state
+from . import account_move_state
