+++ opt_purchase/wizard/wizard_import_helper.py
@@ -45,28 +45,13 @@ class WizardImportHelper(models.TransientModel):
 
 
     # A helper function to set the criteria of what is considered a duplicate
-    def _get_existing_record_searching_domain(self, record):
-        domain = []
+    def _get_existing_record_filtering_domain(self, record):
         if self.name == 'purchase.charge.code':
-            domain = [
-                ('id', '!=', record.id),
-                ('name', '=', record.name),
-                ('project_opt', '=', record.project_opt)
-            ]
+            return lambda r: r.id != record.id and r.name == record.name and r.project_opt == record.project_opt
         elif self.name == 'purchase.account.group':
-            domain = [
-                ('id', '!=', record.id),
-                ('name', '=', record.name),
-                # ('user_id', '=', record.user_id.id),
-                # ('approval_min', '=', record.approval_min)
-            ]
-        # elif self.name == 'purchase.proxy':
-        #     domain = [
-        #         ('id', '!=', record.id),
-        #         ('approver_id', '=', record.approver_id.id),
-        #         ('proxy_id', '=', record.proxy_id.id)
-        #     ]
-        return domain
+            return lambda r: r.id != record.id and r.name == record.name
+        return lambda r: r
+
 
     def do_import(self, model_name, decoded_file, options):
         import_id = self.env['base_import.import'].create({
@@ -75,34 +60,27 @@ class WizardImportHelper(models.TransientModel):
                 'file_type': 'text/csv'
             })
         file_length, rows = import_id._read_file(options)
-        # The first item from data generator is header
         if options.get('has_headers') and rows:
             headers = rows.pop(0)
-            header_types = self._extract_headers_types(headers, rows, options)
+            header_types = import_id._extract_headers_types(headers, rows, options)
         else:
             header_types, headers = {}, []
 
-        # rows_to_import = rows[1:]
         valid_fields = import_id.get_fields_tree(model_name)
         matches = import_id._get_mapping_suggestions(headers, header_types, valid_fields)
 
-        # parsed_header, matches = import_id._match_headers(iter([header]), valid_fields, options)
         matches = {
             header_key[0]: suggestion['field_path']
             for header_key, suggestion in matches.items()
             if suggestion
         }
-        # recognized_fields = [(matches[i] and matches[i][0]) or False for i in range(len(parsed_header))]
-        result = import_id.sudo().execute_import(recognized_fields, parsed_header, options)
-        rids = result.get('ids')
-        if not rids:
+        recognized_fields = [(matches[i] and matches[i][0]) or False for i in range(len(matches))]
+        import_id.sudo().execute_import(recognized_fields, headers, options)
+        if not len(matches):
             raise ValidationError(_('Cannot create/find {} records from the uploaded file.\n'
                                     'Make sure the headers of your file match the technical or functional field names on model {}.\n\n'
                                     'Input Header: {}\n'
-                                    'Mapped Header: {}\n'
-                                    'Error Message: {}'.format(model_name, model_name, parsed_header, recognized_fields,
-                                                               result.get('messages'))))
-        return rids
+                                    'Mapped Header: {}'.format(model_name, model_name, headers, recognized_fields)))
 
 
     def action_set_purchase_levels(self):
@@ -131,7 +109,7 @@ class WizardImportHelper(models.TransientModel):
 
             if user:
                 purchase_level_model = self.env['purchase.level']
-                if not purchase_level_model.search([('name','=', project_id), ('user_id','=', user.id)]):
+                if not purchase_level_model.search([('name', '=', project_id), ('user_id', '=', user.id)]):
                     purchase_level_model.create({
                         'name': project_id,
                         'user_id': user.id,
@@ -153,43 +131,44 @@ class WizardImportHelper(models.TransientModel):
 
     def action_import_records(self):
         self.ensure_one()
+        new_project_ids = []
         if self.name and self.file:
             model_name = self.name
             decoded_file = base64.b64decode(self.file)
-            options = {'quoting': '"', 'separator': ',', 'headers': True}
-            record_lst = self.do_import(model_name, decoded_file, options)
-            # Go through our record lst and unlink any duplicated ones
-            # according to our duplicate domain
+            options = {'quoting': '"', 'separator': ',', 'has_headers': True}
+
+            existing_records = self.env[model_name].with_context(active_test=False).search([])
+            self.do_import(model_name, decoded_file, options)
+            new_records = self.env[model_name].with_context(active_test=False).search([]) - existing_records
+
+            # Remove duplicates by removing old records that match the new ones according to our duplicate domain
             corrected_record_lst = []
             to_unlink = self.env[model_name]
-            new_project_ids = []
             charge_code_project_ids = self.env['purchase.charge.code'].search([('active', '=', True)]).mapped('project_opt')
             purchase_level_project_ids = self.env['purchase.level'].search([]).mapped('name')
 
-            for i in range(len(record_lst)):
-                record_id = record_lst[i]
-                record = self.env[model_name].browse(record_id)
-                existing_record_id = self.env[model_name].with_context(active_test=False).search(self._get_existing_record_searching_domain(record), limit=1)
+            for i in range(len(new_records)):
+                record = new_records[i]
+                existing_record_id = existing_records.filtered(self._get_existing_record_filtering_domain(record))
                 if existing_record_id:
-                    record_id = existing_record_id.id
                     to_unlink |= record
-                    record = existing_record_id
+                    record = existing_record_id[0]
 
-                pid = record.project_opt
-                _logger.info("Importing Project ID: %s" % pid)
+
                 # If a new Project ID is found during import, the process will interrupt and a new dialog will
                 # open that will require the user to select approving users and approval values for the new Project ID
                 # A Product ID is considered new if it is not assigned to any charge code or if it is assigned to a charge Code
                 # but it does not have any purchase levels associated with it
-                if self.name == 'purchase.charge.code' \
-                  and pid not in new_project_ids \
-                  and (pid not in charge_code_project_ids or (pid in charge_code_project_ids and pid not in purchase_level_project_ids)):
-                    new_project_ids.append(pid)
+                if self.name == 'purchase.charge.code':
+                    pid = record.project_opt
+                    if pid not in new_project_ids and (pid not in charge_code_project_ids or (pid in charge_code_project_ids and pid not in purchase_level_project_ids)):
+                        _logger.info("Importing Project ID: %s" % pid)
+                        new_project_ids.append(pid)
 
                 record.sudo().write({
                     'active': True,
                 })
-                corrected_record_lst.append(record_id)
+                corrected_record_lst.append(record.id)
             to_unlink.sudo().unlink()
 
             self.env[model_name].sudo().search([('id', 'not in', corrected_record_lst)]).write({'active': False})
@@ -205,7 +184,7 @@ class WizardImportHelper(models.TransientModel):
                 'res_model': 'wizard.import.helper',
                 'views': [(self.env.ref('opt_purchase.wizard_purchase_levels').id, 'form')],
                 'target': 'new',
-                'context': {'new_project_ids': new_project_ids,'default_project_id': self.project_id},
+                'context': {'new_project_ids': new_project_ids, 'default_project_id': self.project_id},
             }
 
         return {'type': 'ir.actions.act_window_close'}
